Clean-up code, Update Kotlin and Fabric dependency version
