+++ build.gradle
@@ -1,6 +1,6 @@
 buildscript {
 
-    ext.kotlin_version = '1.3.0'
+    ext.kotlin_version = '1.3.11'
 
     repositories {
         google()
@@ -12,7 +12,7 @@ buildscript {
     dependencies {
         classpath 'com.android.tools.build:gradle:3.2.1'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath 'io.fabric.tools:gradle:1.25.4'
+        classpath 'io.fabric.tools:gradle:1.26.1'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
